split single_student into files
