+++ cs251tk/student/__init__.py
@@ -1,2 +1,9 @@
-from .single_student import single_student
+from .analyze import analyze
+from .checkout import checkout_day, checkout_ref
+from .clone import clone
 from .markdownify import markdownify
+from .pull import pull
+from .record import record
+from .remove import remove
+from .reset import reset
+from .stash import stash
