+++ cs251tk/student/remove.py
@@ -1,0 +1,6 @@
+import shutil
+
+
+def remove(student):
+    shutil.rmtree(student)
+
