+++ cs251tk/student/reset.py
@@ -1,0 +1,5 @@
+
+def reset(student, args):
+    with chdir(student):
+        if args['day']:
+            run(['git', 'checkout', 'master', '--quiet', '--force'])
